Move server profiles commands to server profile controller
