+++ WebApi/Controllers/ServersController.cs
@@ -6,21 +6,15 @@ using Sparkle.Application.Common.Interfaces;
 using Sparkle.Application.HubClients.Servers.ServerDeleted;
 using Sparkle.Application.HubClients.Servers.ServerUpdated;
 using Sparkle.Application.Models;
-using Sparkle.Application.Servers.Commands.BanUser;
-using Sparkle.Application.Servers.Commands.ChangeServerProfileDisplayName;
-using Sparkle.Application.Servers.Commands.ChangeServerProfileRoles;
 using Sparkle.Application.Servers.Commands.CreateServer;
 using Sparkle.Application.Servers.Commands.DeleteServer;
 using Sparkle.Application.Servers.Commands.JoinServer;
-using Sparkle.Application.Servers.Commands.KickUser;
 using Sparkle.Application.Servers.Commands.LeaveServer;
-using Sparkle.Application.Servers.Commands.UnbanUser;
 using Sparkle.Application.Servers.Commands.UpdateServer;
 using Sparkle.Application.Servers.Queries.ServerDetails;
 using Sparkle.Application.Servers.Queries.ServersList;
 using Sparkle.Contracts.Servers;
 using Sparkle.WebApi.Attributes;
-using System.ComponentModel;
 
 namespace Sparkle.WebApi.Controllers
 {
@@ -191,132 +185,6 @@ namespace Sparkle.WebApi.Controllers
             Server server = await Mediator.Send(command);
 
             await Mediator.Send(new NotifyServerDeletedQuery { Server = server });
-
-            return NoContent();
-        }
-
-        /// <summary>
-        /// Removes User from the server users list. The User can come back if would have an invitation
-        /// </summary>
-        ///<param name="serverId">Id of the server to kick user from</param>
-        ///<param name="userId">Id of the user to kick from the server</param>
-        /// <returns>No Content if the operation is successful</returns>
-        /// <response code="204">No Content. Operation is successful</response>
-        /// <response code="400">Bad Request. Your request is incorrect</response>
-        /// <response code="401">Unauthorized. The client must be authorized to send this request</response>
-        /// <response code="403">Forbidden. The client has not permissions to perform this action</response>
-        [HttpDelete("{serverId}/profiles/{userId}/kick")]
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-        public async Task<ActionResult> KickUser(string serverId, Guid userId)
-        {
-            KickUserCommand command = new() { ServerId = serverId, UserId = userId };
-            await Mediator.Send(command);
-
-            return NoContent();
-        }
-
-        /// <summary>
-        /// Removes User from the server users list and put him in a black list.
-        /// The User can't come back even if it would have an invitation
-        /// </summary>
-        ///<param name="userId">Id of the user to ban</param>
-        ///<param name="serverId">Id of the server to ban user from</param>
-        /// <returns>No Content if the operation is successful</returns>
-        /// <response code="204">No Content. Operation is successful</response>
-        /// <response code="400">Bad Request. Your request is incorrect</response>
-        /// <response code="401">Unauthorized. The client must be authorized to send this request</response>
-        /// <response code="403">Forbidden. The client has not permissions to perform this action</response>
-        [HttpDelete("{serverId}/profiles/{userId}/ban")]
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-        public async Task<ActionResult> BanUser(string serverId, Guid userId)
-        {
-            BanUserCommand command = new() { ServerId = serverId, UserId = userId };
-            await Mediator.Send(command);
-
-            return NoContent();
-        }
-
-        /// <summary>
-        /// Removes User from the server's black list. Now the user could return if it would have an invitation
-        /// </summary>
-        ///<param name="serverId">Id of the server to unban user from</param>
-        ///<param name="userId">Id of the user to unban</param>
-        /// <returns>No Content if the operation is successful</returns>
-        /// <response code="204">No Content. Operation is successful</response>
-        /// <response code="400">Bad Request. Your request is incorrect</response>
-        /// <response code="401">Unauthorized. The client must be authorized to send this request</response>
-        /// <response code="403">Forbidden. The client has not permissions to perform this action</response>
-        [HttpPost("{serverId}/profiles/{userId}/unban")]
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-        public async Task<ActionResult> UnbanUser(string serverId, Guid userId)
-        {
-            UnbanUserCommand command = new() { ServerId = serverId, UserId = userId };
-            await Mediator.Send(command);
-
-            return NoContent();
-        }
-
-        /// <summary>
-        /// Changes the Display name of the server profile
-        /// </summary>
-        ///<param name="userId">Id of the user to change the display name</param>
-        ///<param name="serverId">Id of the server to change the display name</param>
-        ///<param name="newName">New display name</param>
-        /// <returns>No Content if the operation is successful</returns>
-        /// <response code="204">No Content. Operation is successful</response>
-        /// <response code="400">Bad Request. TYour request is incorrect</response>
-        /// <response code="401">Unauthorized. The client must be authorized to send this request</response>
-        /// <response code="403">Forbidden. The client has not permissions to perform this action</response>
-        [HttpPatch("{serverId}/profile/{userId}/name")]
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-
-        public async Task<ActionResult> ChangeServerProfileDisplayName(
-            [DefaultValue("5f95a3c3d0ddad0017ea9291")] string serverId,
-            Guid userId,
-            string newName)
-        {
-            ChangeServerProfileDisplayNameCommand command = new()
-            {
-                ServerId = serverId,
-                UserId = userId,
-                NewDisplayName = newName
-            };
-            await Mediator.Send(command);
-
-            return NoContent();
-        }
-        /// <summary>
-        /// Changes the set of roles of the give user
-        /// </summary>
-        /// <param name="serverId">Id of the server to change the roles</param>
-        /// <param name="userId">Id of the user to change the roles</param>
-        /// <param name="request">New set of roles</param>
-        /// <returns>No Content if the operation is successful</returns>
-        /// <response code="204">No Content. Operation is successful</response>
-        /// <response code="400">Bad Request. Your request is incorrect</response>
-        /// <response code="401">Unauthorized. The client must be authorized to send this request</response>
-        /// <response code="403">Forbidden. The client has not permissions to perform this action</response>
-        [HttpPatch("{serverId}/profiles/{userId}/roles")]
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-        public async Task<ActionResult> ChangeServerProfileRoles(string serverId, Guid userId, UpdateServerProfileRolesRequest request)
-        {
-            UpdateServerProfileRolesCommand command = _mapper.Map<UpdateServerProfileRolesCommand>((serverId, userId, request));
-            await Mediator.Send(command);
 
             return NoContent();
         }
